Make NavActionManager.getCoverage() return a float instead of double

+++ src/main/java/io/github/omegabird113/cmd_delete/actions/NavActionManager.java
@@ -48,11 +48,11 @@ public class NavActionManager {
         };
     }
 
-    public static double getCoverage(INavMapping mapping) {
+    public static float getCoverage(INavMapping mapping) {
         int total = NavAction.values().length;
         NavAction[] supported = mapping.getPossibleActions();
         int supportCount = supported.length;
 
-        return ((double) supportCount) / total;
+        return ((float) supportCount) / total;
     }
 }
